+++ src/main/java/com/umc/tomorrow/domain/resume/repository/ResumeRepository.java
@@ -10,4 +10,5 @@ import com.umc.tomorrow.domain.resume.entity.Resume;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ResumeRepository extends JpaRepository<Resume, Long> {
+    Resume findByUserId(Long userId);
 }
